Quiet warnings about 64-to-32-bit sized integer types in GlkView.

+++ GlkView/GlkCustomTextSection.m
@@ -24,7 +24,7 @@ NSString* GlkCustomSectionAttributeName = @"GlkCustomSectionAttributeName";
 // = Typesetting =
 
 - (void) placeBaselineAt: (NSPoint) point
-				forGlyph: (int) glyph {
+				forGlyph: (NSInteger) glyph {
 	// Do nothing...
 }
